+++ src/lib/falcor/routes/staff/by-slug.js
@@ -2,7 +2,7 @@ import falcor from 'falcor';
 import _ from 'lodash';
 
 import * as db from 'lib/db';
-import { staffQuery } from './database-calls';
+import { staffQuery } from './database-calls.sql';
 import { has } from 'lib/utilities';
 
 const $ref = falcor.Model.ref;
